syslib: Common kernel and userspace definitions
A trivial library providing a module for constants
and so forth that are shared between the kernel and
userspace.  For example, system call numbers are
defined here.

Both the kernel and ulib depend on syslib; having
this keeps the dependency graph clean, as neither
has to depend directly on the other.

Signed-off-by: Dan Cross <cross@gajendra.net>